f1ml_src/requirements.txt
pandas
numpy
scikit-learn
matplotlib
pyyaml
tqdm
